Moved all the files inside of csci3230 to project root
This was to facilitate the fact that this code is the official code
