+++ Dockerfile
@@ -8,13 +8,11 @@ ENV GRAPHVIZZER_SLACK_AUTHENTICATION_TOKEN="MyMockToken" \
     GRAPHVIZZER_MAXIMUM_DOT_STRING_LENGTH=500 \
     GRAPHVIZZER_SLACK_APP_SECRET="MyMockSecret" \
     GRAPHVIZZER_SLACK_CLIENT_ID="29667068068.63519026177" \
-    GRAPHVIZZER_TEMPORARY_GRAPH_FILE_DIRECTORY="/app/tmp/"
+    GRAPHVIZZER_TEMPORARY_GRAPH_FILE_DIRECTORY="/tmp/"
 
-ADD csci3230-project/ /app/
+ADD dist/ /app/
 
-# Where the temporary files will go, has to be the same as the above GRAPHVIZZER_TEMPORARY_GRAPH_FILE_DIRECTORY
 WORKDIR /app/
-RUN mkdir -p tmp && \
-    npm install --production
+RUN npm install --production
 
 CMD ["npm", "start"]
