+++ app/public/images/graphvizzer_demo.png
