+++ app/temp/README.MD
@@ -1,0 +1,1 @@
+Files are temporary kept here before they are sent to imgur. This directory can be safely deleted.
