+++ app/views/layout.pug
