--- csci3230-project/temp/README.MD
@@ -1,1 +1,0 @@
-Files are temporary kept here before they are sent to imgur
