+++ package.json
@@ -6,13 +6,14 @@
     "test": "test"
   },
   "scripts": {
-    "test": "mocha -R spec spec.js",
-    "test-with-coverage-report": "istanbul cover node_modules/mocha/bin/_mocha -R spec spec.js",
-    "check-style": "jscs server.js"
+    "test": "mocha -R spec test/spec.js",
+    "test-with-coverage-report": "istanbul cover node_modules/mocha/bin/_mocha -R spec test/spec.js",
+    "check-style": "jscs server.js",
+    "start": "node app/server.js"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/ObjectiveTruth/Graphviz-Slack-App.git"
+    "url": "git+https://github.com/ObjectiveTruth/Graphvizzer.git"
   },
   "keywords": [
     "graphvizzer",
@@ -20,11 +21,11 @@
     "dot"
   ],
   "author": "",
-  "license": "ISC",
+  "license": "GPL-3.0",
   "bugs": {
-    "url": "https://github.com/ObjectiveTruth/Graphviz-Slack-App/issues"
+    "url": "https://github.com/ObjectiveTruth/Graphvizzer/issues"
   },
-  "homepage": "https://github.com/ObjectiveTruth/Graphviz-Slack-App#readme",
+  "homepage": "https://graphvizzer.objectivetruth.ca",
   "dependencies": {
     "body-parser": "^1.15.2",
     "child_process": "^1.0.2",
